WIP: Python 12 and 13 support. - Add classifiers. - Move `resolve_flavor`.
