+++ ci/deps/.gitignore
@@ -4,5 +4,7 @@
 # Except
 !actions-310.yml
 !actions-311.yml
+!actions-312.yml
+!actions-313.yml
 
 !.gitignore
