+++ docs/templates/usage.md.jinja
@@ -2,7 +2,6 @@ To harness the toolkit's capabilities, you should import the necessary packages:
 
 ```python
 import re
-# and/or
 import re2
 import regex_toolkit as rtk
 ```
